service/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0
websockets==12.0

# Image processing
Pillow==10.2.0
# For production, replace with pillow-simd for 4-6x faster Lanczos:
# pip uninstall pillow && CC="cc -mavx2" pip install pillow-simd

# Background removal (local models)
rembg==2.0.50
onnxruntime==1.16.3  # CPU version; use onnxruntime-gpu for GPU

# Runware SDK
runware==0.2.0

# HTTP client for Kie.ai
requests==2.31.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
PyYAML==6.0.1

# Utilities
aiofiles==23.2.1
python-magic==0.4.27  # MIME type detection

# Logging
structlog==24.1.0

# Optional: Metrics
# prometheus-client==0.19.0
